backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
pytest
pytest-asyncio
elevenlabs
google-genai
